Изменить тип индекса с bigint на int

+++ notifications_service/src/main/java/com/subOne/notifications_service/dto/notification/response/ResponseNotificationDto.java
@@ -5,5 +5,5 @@ import com.subOne.notification.NotificationType;
 
 import java.time.OffsetDateTime;
 
-public record ResponseNotificationDto(Long id, NotificationTargetType notificationTargetType, NotificationType notificationType, Long targetId, OffsetDateTime createdAt) {
+public record ResponseNotificationDto(Integer id, NotificationTargetType notificationTargetType, NotificationType notificationType, Integer targetId, OffsetDateTime createdAt) {
 }
