+++ subscriptions_service/src/main/java/com/subOne/subscriptions_service/dto/analytic_subscription/SubscriptionLastDatePaymentIdAndGroupIdDto.java
@@ -4,5 +4,5 @@ package com.subOne.subscriptions_service.dto.analytic_subscription;
 import java.time.LocalDate;
 
 
-public record SubscriptionLastDatePaymentIdAndGroupIdDto(Long subscriptionId, Long groupId, String paymentPeriod, LocalDate datePaid) {
+public record SubscriptionLastDatePaymentIdAndGroupIdDto(Integer subscriptionId, Integer groupId, String paymentPeriod, LocalDate datePaid) {
 }
