+++ user_service/src/main/java/com/subOne/user_service/dto/user/request/RequestGroupOwnershipChangesDto.java
@@ -2,5 +2,5 @@ package com.subOne.user_service.dto.user.request;
 
 import java.util.UUID;
 
-public record RequestGroupOwnershipChangesDto(UUID userId, Long groupId) {
+public record RequestGroupOwnershipChangesDto(UUID userId, Integer groupId) {
 }
